third_party/blink/web_tests/external/wpt/html/canvas/element/wide-gamut-canvas/2d.color.space.p3.strokeText-expected.txt

This is a testharness.js-based test.
FAIL Test if strokeText can be used with a solid display-p3 color assert_array_equals: expected property 0 to be 255 but got 234 (expected array [255, 0, 0, 255] got [234, 51, 35, 255])
Harness: the test ran to completion.
